fix husky
