+++ .husky/pre-push
@@ -1,4 +1,6 @@
 #!/bin/sh
 . "$(dirname "$0")/_/husky.sh"
 
-echo "check commit"
+npm run test
+
+npm run build
