+++ src/index.html
@@ -12,11 +12,12 @@
       integrity="sha384-Gn5384xqQ1aoWXA+058RXPxPg6fy4IWvTNh0E263XmFcJlSAwiGgFAW/dAiS6JXm"
       crossorigin="anonymous"
     />
-    <title>ЦЕЛИ</title>
+    <title>ToDoList</title>
   </head>
   <body>
     <div class="container">
-      <div class="h4 mb-5">Список поставленных задач</div>
+      <h1>Список поставленных задач</h1>
+      <div class="h4 mb-5"></div>
       <div class="card mb-4">
         <ul id="tasksList" class="list-group list-group-flush"></ul>
       </div>
