+++ src/redux/reducers.ts
@@ -1,35 +1,23 @@
-/* eslint-disable no-param-reassign */
 import { createSlice, PayloadAction } from "@reduxjs/toolkit";
-import { Task } from "./actions";
+import { Task, TasksState } from "./actions";
 
-interface TasksState {
-  tasks: Task[];
-}
-
-const initialState: TasksState = {
-  tasks: [],
-};
+const initialState: TasksState = [];
 
 const tasksSlice = createSlice({
   name: "tasks",
   initialState,
   reducers: {
-    addTask: (state, action: PayloadAction<Task>) => {
-      state.tasks.push(action.payload);
-    },
-    deleteTask: (state, action: PayloadAction<number>) => {
-      state.tasks = state.tasks.filter((task) => task.id !== action.payload);
-    },
-    toggleTask: (state, action: PayloadAction<number>) => {
-      const foundTaskIndex = state.tasks.findIndex(
-        (t) => t.id === action.payload,
-      );
-      if (foundTaskIndex !== -1) {
-        state.tasks[foundTaskIndex].done = !state.tasks[foundTaskIndex].done;
-      }
-    },
+    initTasks: (_, action: PayloadAction<Task[]>) => action.payload,
+    addTask: (state, action: PayloadAction<Task>) => [...state, action.payload],
+    deleteTask: (state, action: PayloadAction<number>) =>
+      state.filter((task) => task.id !== action.payload),
+    toggleTask: (state, action: PayloadAction<number>) =>
+      state.map((task) =>
+        task.id === action.payload ? { ...task, done: !task.done } : task,
+      ),
   },
 });
 
-export const { addTask, deleteTask, toggleTask } = tasksSlice.actions;
+export const { initTasks, addTask, deleteTask, toggleTask } =
+  tasksSlice.actions;
 export default tasksSlice.reducer;
